rer-python/pyproject.toml
[project]
name = ""
version = ""
description = "API calls. Fill the config with endpoints and parameters, run the main.py"
authors = [
    {name = "ReRubis", email = "slesarenkod@gmail.com"},
]
dependencies = [
    "asyncio>=3.4.3",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
    "aiohttp>=3.9.0b1",
]
requires-python = ">=3.12"
license = {text = "MIT"}

